Move activities controller and views to be scoped to schools (as per interventions)
