+++ app/views/schools/activities/_form.html.erb
