+++ app/views/schools/activities/completed.html.erb
