+++ app/views/schools/activities/duplicate_warning.html.erb
