+++ app/views/schools/activities/edit.html.erb
