+++ app/views/schools/activities/new.html.erb
