+++ app/views/schools/activities/show.html.erb
@@ -8,25 +8,25 @@
         <%= link_to @school.name, school_path(@school) %>, <%= @activity.happened_on.strftime('%A, %d %B %Y') %>
       </span>
       <br>
-      <span class="badge badge-success"
+      <span class="badge text-bg-success"
             data-toggle="tooltip"
             data-placement="bottom" title="<%= t('activities.show.score_awarded_for_this_activity') %>">
             <%= t('common.points_html', count: @activity.points) %>
       </span>
       <% if @activity.pupil_count %>
-        <span class="badge badge-info"
+        <span class="badge text-bg-info"
               data-toggle="tooltip"
               data-placement="bottom" title="<%= t('interventions.show.pupils_involved_in_this_activity') %>">
               <%= t('common.pupil_count', count: @activity.pupil_count) %>
         </span>
       <% end %>
-      <span class="badge badge-secondary"
+      <span class="badge text-bg-secondary"
             data-toggle="tooltip"
             data-placement="bottom"
             title="<%= t('activities.show.category_of_activity') %>">
             <%= @activity.activity_type.activity_category.name %>
       </span>
-      <span class="badge badge-primary"
+      <span class="badge text-bg-primary"
             data-toggle="tooltip"
             data-placement="bottom" title="<%= t('activities.show.key_stages') %>">
             <%= @activity.activity_type.key_stage_list %>
